Stage 1h pivot B15 (inner): rsync src/intersim2/ (missed in pivot scope)
Audit of still-diff'd files between ours and MICRO 2025 discovered
src/intersim2/ was never rsync'd during the original pivot.  14 files
modified (-21 net lines).

Not the cause of the bitset OOR crash (which was ABI staleness — see
feedback memory), but completes the pivot scope properly.  intersim2
implements the ICNT (interconnect) simulation layer; same-version
alignment with MICRO 2025 is prerequisite for future modeling work.

+++ src/intersim2/flit.hpp
@@ -58,7 +58,7 @@ public:
   int  atime;
 
   unsigned long long  id;
-  unsigned long long  pid;
+  int  pid;
 
   bool record;
